Get posted jobs

+++ src/main/java/Propath/service/CompanyService.java
@@ -1,6 +1,7 @@
 package Propath.service;
 
 import Propath.dto.CompanyDto;
+import Propath.dto.PostJobDto;
 import Propath.model.Company;
 
 import java.util.List;
@@ -14,5 +15,8 @@ public interface CompanyService {
 
     CompanyDto updateCompanyStatus(Long id,CompanyDto updatedCompany);
 
-    CompanyDto getCompanyByUserId(Long userId);
+    CompanyDto getCompanyByUserId(int userId);
+
+
+
 }
